[dijkstra][include] upgrading dijkstra code
